change basicmodule to use Base POM module as a parent module
