+++ omod/src/main/resources/config.xml
@@ -6,23 +6,23 @@
 	<id>${project.parent.artifactId}</id>
 	<name>${project.parent.name}</name>
 	<version>${project.parent.version}</version>
-	<package>${project.parent.groupId}.${project.parent.artifactId}</package>
+	<package>${MODULE_PACKAGE}</package>
 	<author>Ben Wolfe</author>
 	<description>
 			${project.parent.description}
 	</description>
 
-	<activator>${project.parent.groupId}.${project.parent.artifactId}.BasicModuleActivator</activator>
+	<activator>${MODULE_PACKAGE}.BasicModuleActivator</activator>
 	
 	<!-- <updateURL>https://modules.openmrs.org/modules/download/@MODULE_ID@/update.rdf</updateURL> -->
 	<!-- /Base Module Properties -->
 	
-	<require_version>${openMRSVersion}</require_version>
+	<require_version>${openmrs.version}</require_version>
 	
 	<!-- Extensions -->
 	<extension>
 		<point>org.openmrs.admin.list</point>
-		<class>@MODULE_PACKAGE@.extension.html.AdminList</class>
+		<class>${MODULE_PACKAGE}.extension.html.AdminList</class>
 	</extension>
